# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "FastAPI",
      "version": "0.1.0"
    },
    "paths": {
...
      "/requirements/": {
        "get": {
          "tags": [
            "requirements"
          ],
          "summary": "Get Requirements",
          "operationId": "get_requirements_requirements__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/RequirementResponse"
                    },
                    "type": "array",
                    "title": "Response Get Requirements Requirements  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "requirements"
          ],
          "summary": "Create Requirement",
          "operationId": "create_requirement_requirements__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RequirementResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/requirements/{requirement_id}": {
        "get": {
          "tags": [
            "requirements"
          ],
          "summary": "Get Requirement",
          "operationId": "get_requirement_requirements__requirement_id__get",
          "parameters": [
            {
              "name": "requirement_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Requirement Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RequirementResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "requirements"
          ],
          "summary": "Delete Requirement",
          "operationId": "delete_requirement_requirements__requirement_id__delete",
          "parameters": [
            {
              "name": "requirement_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Requirement Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/requirements/with-articles": {
        "post": {
          "tags": [
            "requirements"
          ],
          "summary": "Create Requirement With Articles",
          "description": "Create a requirement with its associated articles",
          "operationId": "create_requirement_with_articles_requirements_with_articles_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementWithArticlesCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RequirementResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/suppliers/": {
        "get": {
          "tags": [
            "suppliers"
          ],
          "summary": "Get Suppliers",
          "operationId": "get_suppliers_suppliers__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/SupplierResponse"
                    },
                    "type": "array",
                    "title": "Response Get Suppliers Suppliers  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "suppliers"
          ],
          "summary": "Create Supplier",
          "operationId": "create_supplier_suppliers__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SupplierResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/suppliers/{supplier_id}": {
        "get": {
          "tags": [
            "suppliers"
          ],
          "summary": "Get Supplier",
          "operationId": "get_supplier_suppliers__supplier_id__get",
          "parameters": [
            {
              "name": "supplier_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Supplier Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SupplierResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "suppliers"
          ],
          "summary": "Delete Supplier",
          "operationId": "delete_supplier_suppliers__supplier_id__delete",
          "parameters": [
            {
              "name": "supplier_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Supplier Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "suppliers"
          ],
          "summary": "Update Supplier",
          "operationId": "update_supplier_suppliers__supplier_id__put",
          "parameters": [
            {
              "name": "supplier_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Supplier Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SupplierResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "ArticleCreateWithoutRequirement": {
          "properties": {
            "quantity": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ],
              "title": "Quantity"
            },
            "unit": {
              "type": "string",
              "title": "Unit"
            },
            "brand": {
              "type": "string",
              "title": "Brand"
            },
            "model": {
              "type": "string",
              "title": "Model"
            },
            "dimensions": {
              "type": "string",
              "title": "Dimensions"
            },
            "state_id": {
              "type": "integer",
              "title": "State Id"
            },
            "notes": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notes"
            }
          },
          "type": "object",
          "required": [
            "quantity",
            "unit",
            "brand",
            "model",
            "dimensions",
            "state_id"
          ],
          "title": "ArticleCreateWithoutRequirement",
          "description": "Model for creating an article without requirement_id"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "RequirementCreate": {
          "properties": {
            "project_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            },
            "request_date": {
              "type": "string",
              "format": "date-time",
              "title": "Request Date"
            },
            "requested_by": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requested By"
            },
            "state_id": {
              "type": "integer",
              "title": "State Id"
            },
            "closing_date": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Closing Date"
            }
          },
          "type": "object",
          "required": [
            "state_id"
          ],
          "title": "RequirementCreate"
        },
        "RequirementResponse": {
          "properties": {
            "project_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            },
            "request_date": {
              "type": "string",
              "format": "date-time",
              "title": "Request Date"
            },
            "requested_by": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requested By"
            },
            "state_id": {
              "type": "integer",
              "title": "State Id"
            },
            "closing_date": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Closing Date"
            },
            "id": {
              "type": "integer",
              "title": "Id"
            }
          },
          "type": "object",
          "required": [
            "state_id",
            "id"
          ],
          "title": "RequirementResponse"
        },
...
        "RequirementWithArticlesCreate": {
          "properties": {
            "requirement": {
              "$ref": "#/components/schemas/RequirementCreate"
            },
            "articles": {
              "items": {
                "$ref": "#/components/schemas/ArticleCreateWithoutRequirement"
              },
              "type": "array",
              "title": "Articles"
            }
          },
          "type": "object",
          "required": [
            "requirement",
            "articles"
          ],
          "title": "RequirementWithArticlesCreate",
          "description": "Model for creating a requirement with its articles"
        },
        "SupplierCreate": {
          "properties": {
            "name": {
              "type": "string",
              "title": "Name"
            },
            "rfc": {
              "type": "string",
              "title": "Rfc"
            },
            "address_id": {
              "type": "integer",
              "title": "Address Id"
            },
            "bank_details": {
              "type": "string",
              "title": "Bank Details"
            },
            "delivery_time": {
              "type": "string",
              "title": "Delivery Time"
            },
            "payment_condition_id": {
              "type": "integer",
              "title": "Payment Condition Id"
            },
            "currency": {
              "type": "string",
              "title": "Currency"
            },
            "notes": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notes"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            }
          },
          "type": "object",
          "required": [
            "name",
            "rfc",
            "address_id",
            "bank_details",
            "delivery_time",
            "payment_condition_id",
            "currency"
          ],
          "title": "SupplierCreate"
        },
        "SupplierResponse": {
          "properties": {
            "name": {
              "type": "string",
              "title": "Name"
            },
            "rfc": {
              "type": "string",
              "title": "Rfc"
            },
            "address_id": {
              "type": "integer",
              "title": "Address Id"
            },
            "bank_details": {
              "type": "string",
              "title": "Bank Details"
            },
            "delivery_time": {
              "type": "string",
              "title": "Delivery Time"
            },
            "payment_condition_id": {
              "type": "integer",
              "title": "Payment Condition Id"
            },
            "currency": {
              "type": "string",
              "title": "Currency"
            },
            "notes": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notes"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            },
            "id": {
              "type": "integer",
              "title": "Id"
            }
          },
          "type": "object",
          "required": [
            "name",
            "rfc",
            "address_id",
            "bank_details",
            "delivery_time",
            "payment_condition_id",
            "currency",
            "id"
          ],
          "title": "SupplierResponse"
        },
        "SupplierUpdate": {
          "properties": {
            "name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            },
            "rfc": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rfc"
            },
            "address_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Address Id"
            },
            "bank_details": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bank Details"
            },
            "delivery_time": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Delivery Time"
            },
            "payment_condition_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Payment Condition Id"
            },
            "currency": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Currency"
            },
            "notes": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notes"
            }
          },
          "type": "object",
          "title": "SupplierUpdate"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      }
    }
  }